backend files changes

+++ Fuel_Station/src/main/java/Fuel_Station/Fuel_Station/Service/TransactionServiceImpl.java
@@ -1,12 +1,12 @@
 package Fuel_Station.Fuel_Station.Service;
 
-import Fuel_Station.Fuel_Station.Entity.FuelStation;
-import Fuel_Station.Fuel_Station.Entity.Transaction;
+import Fuel_Station.Fuel_Station.Entity.*;
 import Fuel_Station.Fuel_Station.Repository.TransactionRepository;
 import Fuel_Station.Fuel_Station.dto.request.TransactionRequest;
 import Fuel_Station.Fuel_Station.dto.response.FuelStationResponse;
 import Fuel_Station.Fuel_Station.dto.response.MessageResponse;
 import Fuel_Station.Fuel_Station.dto.response.TransactionResponse;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.pulsar.PulsarProperties;
 import org.springframework.http.ResponseEntity;
@@ -22,6 +22,12 @@ public class TransactionServiceImpl implements TransactionService{
     private TransactionRepository transactionRepository;
     @Autowired
     private FuelStationService fuelStationService;
+    @Autowired
+    private EmployeeService employeeService;
+    @Autowired
+    private FuelService fuelService;
+    @Autowired
+    private VehicleService vehicleService;
 
     public Transaction getById(Long transactionId){
         return transactionRepository.getByTransactionId(transactionId).orElse(null);
@@ -33,12 +39,11 @@ public class TransactionServiceImpl implements TransactionService{
         List<TransactionResponse> responses=new ArrayList<>();
         for (Transaction transaction:transactionList){
             TransactionResponse response = new TransactionResponse(
-            transaction.getTransactionId(),
-            transaction.getDateTime(),
-            transaction.getQuantity()
-        
+                    transaction.getTransactionId(),
+                    transaction.getQuantity(),
+                    transaction.getDateTime()
             );
-        responses.add(response);
+            responses.add(response);
         }
             return ResponseEntity.ok().body(
                     new MessageResponse<>(
@@ -53,14 +58,14 @@ public class TransactionServiceImpl implements TransactionService{
     public ResponseEntity<?> getTransactionById(Long transactionId) {
         Optional<Transaction> OptionalTransaction = transactionRepository.findById(transactionId);
         if (OptionalTransaction.isEmpty()) {
-            
-        return ResponseEntity.ok().body(
-            new MessageResponse<>(
-                    400,
-                    "Transaction id not found",
-                    null
-            )
-    );}
+            return ResponseEntity.ok().body(
+                new MessageResponse<>(
+                        400,
+                        "Transaction id not found",
+                        null
+                )
+            );
+        }
             Transaction transaction=OptionalTransaction.get();
             TransactionResponse response=new TransactionResponse(
                 transaction.getTransactionId(),
@@ -75,37 +80,80 @@ public class TransactionServiceImpl implements TransactionService{
 
     @Override
     public ResponseEntity<?> addTransaction(TransactionRequest transactionRequest) {
-        Transaction transaction =new TransactionResponse(
-
-                transactionRequest.getDateTime(),
-                transactionRequest.getQuantity());
-               
+        FuelStation fuelStation = fuelStationService.getById(transactionRequest.getFuelStationId());
+        if(fuelStation == null){
+            return ResponseEntity.ok().body(
+                    new MessageResponse<>(
+                            400,
+                            "Fuel station not found",
+                            null
+                    )
+            );
+        }
+        Fuel fuel = fuelService.getById(transactionRequest.getFuelId());
+        if(fuel == null){
+            return ResponseEntity.ok().body(
+                    new MessageResponse<>(
+                            400,
+                            "Fuel not found",
+                            null
+                    )
+            );
+        }
+        Employee employee = employeeService.getById(transactionRequest.getEmployeeId());
+        if(employee == null){
+            return ResponseEntity.ok().body(
+                    new MessageResponse<>(
+                            400,
+                            "Employee not found",
+                            null
+                    )
+            );
+        }
+        Vehicle vehicle = vehicleService.getById(transactionRequest.getVehicleId());
+        if(vehicle == null){
+            return ResponseEntity.ok().body(
+                    new MessageResponse<>(
+                            400,
+                            "Vehicle not found",
+                            null
+                    )
+            );
+        }
+        Transaction transaction = new Transaction(
+                fuelStation,
+                fuel,
+                vehicle,
+                employee,
+                transactionRequest.getQuantity(),
+                transactionRequest.getDateTime()
+        );
         transactionRepository.save(transaction);
         return ResponseEntity.ok().body(
                 new MessageResponse<>(
                         200,
-                        "Fuel station added successfully",
+                        "Transaction added Successfully",
                         null
                 )
         );
     }
 
     @Override
-    public ResponseEntity<?> updateTransaction(Long transactionId, TransactionRequest transaction) {
-        Transaction existingTransaction =  transactionRepository.getByTransactionId(transactionId);
-if (existingTransaction==null) {
-    return ResponseEntity.ok().body(
-        new MessageResponse<>(
-                400,
-                "Fuel station not found",
-                null
-        )
-);
-    
-}
-        existingTransaction.setQuantity(transaction.getQuantity());
-        existingTransaction.setDateTime(transaction.getDateTime());
-         transactionRepository.save(existingTransaction);
+    public ResponseEntity<?> updateTransaction(Long transactionId, TransactionRequest transactionRequest) {
+        Optional<Transaction> optionalTransaction =  transactionRepository.getByTransactionId(transactionId);
+        if (optionalTransaction.isEmpty()) {
+         return ResponseEntity.ok().body(
+            new MessageResponse<>(
+                    400,
+                    "Transaction not found",
+                    null
+            )
+        );
+    }
+        Transaction transaction = optionalTransaction.get();
+        transaction.setQuantity(transaction.getQuantity());
+        transaction.setDateTime(transaction.getDateTime());
+         transactionRepository.save(transaction);
          return ResponseEntity.ok().body( 
              new MessageResponse<>(
             200,
@@ -116,7 +164,18 @@ if (existingTransaction==null) {
     
 
     @Override
+    @Transactional
     public ResponseEntity<?> deleteTransaction(Long transactionId) {
+        Optional<Transaction> transaction = transactionRepository.getByTransactionId(transactionId);
+        if(transaction.isEmpty()){
+            return ResponseEntity.ok().body(
+                    new MessageResponse<>(
+                            400,
+                            "Transaction not found",
+                            null
+                    )
+            );
+        }
         transactionRepository.deleteById(transactionId);
         return ResponseEntity.ok().body(
             new MessageResponse<>(
@@ -128,24 +187,24 @@ if (existingTransaction==null) {
 
     public  ResponseEntity<?>  getTransactionsByStationId(Long stationId) {
         FuelStation fuelStation = fuelStationService.getById(stationId);
-        if(fuelStation != null){
+        if(fuelStation == null){
             return ResponseEntity.ok().body(
                     new MessageResponse<>(
                             400,
-                            "Owner not found",
+                            "Station not found",
                             null
                     )
             );
         }
-          List<Transaction> transactionList=  transactionRepository.findByFuelStation(fuelStation);
+        List<Transaction> transactionList=  transactionRepository.findByFuelStation(fuelStation);
         List<TransactionResponse> responses=new ArrayList<>();
         for(Transaction transaction:transactionList){
             TransactionResponse response=new TransactionResponse(
-            transaction.getDateTime(),
-            transaction.getTransactionId(),
-            transaction.getQuantity()
+                    transaction.getTransactionId(),
+                    transaction.getQuantity(),
+                    transaction.getDateTime()
             );
-responses.add(response);
+            responses.add(response);
         }
         return ResponseEntity.ok().body(
             new MessageResponse<>(
